feat: CSP support
